Redesign macros and make hooks return pointers
This commit involves two major breaking changes to the crate interface: it
makes the `use_` methods return pointers, and it redesigns the interface
on the shareable_struct! macro to allow adding substructs.

The documentation has been updated with exxamples of the new syntax.

Both of these changes can be implemented without rolling over the major
version, but the eventual goal is to roll over when dioxus does and this
is the version of the features that will most likely go in that release.

In the meantime, after more testing I'll port these changes to version 1.2 in a
way that is SemVer friendly.