Review and comment unsafe code.
This update reviews the unsafe code and adds # SAFETY: comments where
necessary.

Closes #1.

Code accessing the a shared static was deemed to have insufficiently robust safety
guarantees; if the shared value was initialized by two concurrent
processes, it could lead to those two processes holding pointers to
different data.

To fix the race condition, a `Mutex` was added as a wrapper on the
container, and we switched from RefCell to parking_lot::RwLock.
As a result shared data must now be `Send + Sync`.

BREAKING CHANGE: Shared data must now be `Send + Sync`