refactor: add `ArcMap` type and use it in `Shareable`.
This update adds a new type `ArcMap<T>` which uses the guarantees of
`Arc<U>` to guard pointers to subfields of `U` and then erases the type
`U`. The idea is to allow sharing subfields of a structure individually
without needing them to be stored behind separate pointers; hopefully,
this will allow for more complicated shared structures to be both
ergonomic and efficient.

`Shareable` now wraps its data in an `ArcMap` rather than an `Arc`,
which should allow for a rewrite of `shareable_struct!` to use one `Arc`
with several `Link` values inside.