refactor: switch to using a single pointer for structs
This commit updates the storage for `shareable_struct!` so that the
struct is defined as a single pointer to a struct containing multiple
`Link<_>`s (one for each field) rather than a separate pointer and allocation
for each field. Full changes (using a single pointer for the struct
when it is used as well) is still WIP.

fix: This commit also changes the syntax on several methods to use
`dioxus_core::Scope` rather than `&dioxus_core::Scope`.