refactor: modify `struct_assoc_type!` to improve compiler type inference
This commit splits the `AssocType` trait into 3 separate traits,
`AssocSubstruct`, `AssocField`, `AssocAction` which require that the
association implements the appropriate traits for a substruct, field, or
action respectively. The goal is to guarantee that the compiler can
always infer these relationships correctly.

This was probably unneccessary (AFAIK there is no reasonable case in
which the types used here are not concrete), however, it potentially
opens up the possibility of better error messages with a couple more
modifications.