refactor: remove unnecessary generic trait bounds
This commit simplifies type resolution by removing some unneccessary
trait bounds.

BUGS: it was discovered in the process of going over the current state
of the code that while this branch compiles, the field accessor methods
on shareable structures were stubbed and never written. Unfortunately,
due to a compiler bug, a slight refactor is going to be needed to write
those methods the way I initially intended to.